Fix big album art size and colorize music visualizer:
Always scale the album art bitmap to the size of the ImageView.
If the album art image is smaller than the ImageView, it won't be scaled
up to fit the ImageView, which results in a bigger music visualizer
than the album art image.

Use Android support library 'Palette' class to extract a color from the album
art image to colorize the music visualizer, if the media meta data doesn't
contain a album art image, use the default visualizer color.