Add landscape layout and improve portrait layout:
- Bottom sheet:
  - Change big album art image size from 250dp to 300dp
  - Move content to the top of the top panel on expand slide
  - Move back content to the bottom of the top panel on collapse slide
  - Change expand/collapse slide animation:
    - 0% - 50% expanded: Top panel alpha,
     (0% expanded: fully visible, 50% expanded: fully invisible)
    - 50% - 100 expanded: Content translationY,
     (50% expanded: translationY = 0dp, 100%  expanded: translationY = -64dp)
  - Portrait:
    - Remove unneeded LinearLayout

NOTE:
On portrait orientation, there is no need to move the
bottom sheet content to the top, even with the bigger album art image, there is enough
height to show all the content on the screen, (at least on the Nexus 5X).
But on landscape orientation, the bigger album art image doesn't fit into
the screen height. I decided to move the bottom sheet content to the top
on portrait orientation as well, in my opinion it looks much better,
when the album art image is placed to the bottom of the action bar,
instead of a 64dp blank space between album art image and action bar.