Bottom sheet: Play/pause button big add song position seeking and more:
Import 'PlayPauseButton' and 'PlayPauseProgressButton' from CyanogenMod/LineageOS:
PlayPauseButton:
https://github.com/LineageOS/android_packages_apps_Eleven/blob/lineage-16.0/src/org/lineageos/eleven/widgets/PlayPauseButton.java
PlayPauseProgressButton:
https://github.com/LineageOS/android_packages_apps_Eleven/blob/lineage-16.0/src/org/lineageos/eleven/widgets/PlayPauseProgressButton.java

- Handle the case the current song was played completely:
  - Set seek bar and play/pause progress button position to '0'
  - Set play/pause button icons to 'play'
- Remove unneeded drawables
- Small clean up

Fix bottom sheet expand/collapse issues:
Currently the bootom sheet (top) bar, (main screen bottom bar)
slides out to the top of the bottom sheet when expanding,
and slides in from the top of the bottom sheet when collapsing.
This will result in massive lags when expanding/collapsing the bottom sheet,
and sometimes the bottom sheet won't collapse completely.
I couldn't find a way to fix this issue.

I decided to made some changes to the layouts, so there is
no need to slide in/out the 'top/bottom' bar anymore,
(also some general changes/additions not related to the issue):

- Bootom sheet (top) bar, (main screen bottom bar):
  - Move to the main content layout.
  - Place the play/pause button to the right
- Bottom sheet
  - Fully transparent when collapsed,
    (the bottom bar is behind the bottom sheet,
    so the bottom bar is visible when collapsed)
  - Fully opaque when not collapsed
    (this will hide the bottom bar)
  - Add a drag icon to the top

TO-DO:
- Clean up
- Stop music visualizer when bottom sheet is collapsed
- Stop 'PlayPauseProgressButton' progress updates when bottom sheet is collapsed
- Remove the seek bar
- Place current song playtime and -remaining time to the left/right of the 'PlayPauseProgressButton'
- Replace 'simple' play/pause button on the bottom bar with 'PlayPauseProgressButton',
  (with disabled song position seeking)
- Improve landscape layout