Add service to support background playback:
Use a Service to be able to play music in the background.
The service will run in the foreground when currently playing music,
and run in the background when playback is paused.
Use the 'default' music notification as foreground service notification,
(notification won't be dismissable).
When the playback is paused, the notification will be dismissable.

MediaSession:
Add the album art bitmap to the MediaSession,
also add the current playback state, to show the album art as lock screen background,
(should be optional later).
NOTE:
This only a very basic MediaSession implementation,
and for now, only album art on lock screen is supportet,
no playback control via headset, etc. ...

TO-DO:
Handle incomming calls, notifications, audio focus changes, ...
Add support for 'external' playback control (headset, ...)
Add more infos to the MediaSession (song title, artist, ...)