Improve service lifecycle and more:
As in the current service implementation the service won't be stopped,
(at least as long as the system doesn't kill the service),
stop the service when:
(all conditions must be met)
- No client is bound to the service
  (means, the main screen isn't visible,
  e.g. settings screen is visible, app is closed/in background, ...)
- The playback is in idle state,
  (paused, or wasn't started)
- The idle state takes longer than 5 minutes

Save the current song (info) to the shared preferences,
set the saved playback position to '0'
when a song was loaded, (click on the song list),
and reload the song (info) from the shared preferences
when the service was created, so the last (loaded) song will always be 'available'.

Save the current playback position to the shared preferences
when the service was stopped,
(the service methode 'onDestroy()' has to be called,
which isn't guaranteed to happen in all cases),
and restore the saved playback position from the shared preferences
when the service was created.

Play/pause progress button (small)
- Fix enabled state when the bottom sheet was closed
- Apply disabled state to the play/pause button:
  - Disables click events
  - Show the play/pause icon disabled

TO-DO:
Don't start the service when loading a song,
only start the service when playback starts.
Add a stop/reset function, which will stop the service,
release the media player, ...